+++ demos/default/pages/getServerSideProps/[id].js
@@ -25,9 +25,7 @@ const Show = ({ errorCode, show, env }) => {
 
       <hr />
 
-      <Link href="/">
-        <a>Go back home</a>
-      </Link>
+      <Link href="/">Go back home</Link>
     </div>
   )
 }