+++ demos/default/pages/getStaticProps/withFallback/[id].js
@@ -23,9 +23,7 @@ const Show = ({ show, time }) => {
       <p>Rendered at {time} </p>
       <hr />
 
-      <Link href="/">
-        <a>Go back home</a>
-      </Link>
+      <Link href="/">Go back home</Link>
     </div>
   )
 }