+++ demos/default/pages/static/[id].js
@@ -23,9 +23,7 @@ const StaticWithID = (props) => (
 
     <hr />
 
-    <Link href="/">
-      <a>Go back home</a>
-    </Link>
+    <Link href="/">Go back home</Link>
   </div>
 )