+++ packages/runtime/src/helpers/edge.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines */
 import { promises as fs, existsSync } from 'fs'
 import { resolve, join } from 'path'
 
@@ -12,6 +11,7 @@ import { outdent } from 'outdent'
 
 import { getRequiredServerFiles, NextConfig } from './config'
 import { makeLocaleOptional, stripLookahead } from './matchers'
+import { RoutesManifest } from './types'
 
 // This is the format as of next@12.2
 interface EdgeFunctionDefinitionV1 {
@@ -22,6 +22,10 @@ interface EdgeFunctionDefinitionV1 {
   regexp: string
 }
 
+interface AssetRef {
+  name: string
+  filePath: string
+}
 export interface MiddlewareMatcher {
   regexp: string
   locale?: false
@@ -35,6 +39,8 @@ interface EdgeFunctionDefinitionV2 {
   name: string
   page: string
   matchers: MiddlewareMatcher[]
+  wasm?: AssetRef[]
+  assets?: AssetRef[]
 }
 
 type EdgeFunctionDefinition = EdgeFunctionDefinitionV1 | EdgeFunctionDefinitionV2
@@ -46,23 +52,29 @@ export interface FunctionManifest {
         function: string
         name?: string
         path: string
+        cache?: 'manual'
       }
     | {
         function: string
         name?: string
         pattern: string
+        cache?: 'manual'
       }
   >
   import_map?: string
 }
 
-export const loadMiddlewareManifest = (netlifyConfig: NetlifyConfig): Promise<MiddlewareManifest | null> => {
-  const middlewarePath = resolve(netlifyConfig.build.publish, 'server', 'middleware-manifest.json')
-  if (!existsSync(middlewarePath)) {
-    return null
+const maybeLoadJson = <T>(path: string): Promise<T> | null => {
+  if (existsSync(path)) {
+    return readJson(path)
   }
-  return readJson(middlewarePath)
 }
+
+export const loadMiddlewareManifest = (netlifyConfig: NetlifyConfig): Promise<MiddlewareManifest | null> =>
+  maybeLoadJson(resolve(netlifyConfig.build.publish, 'server', 'middleware-manifest.json'))
+
+export const loadAppPathRoutesManifest = (netlifyConfig: NetlifyConfig): Promise<Record<string, string> | null> =>
+  maybeLoadJson(resolve(netlifyConfig.build.publish, 'app-path-routes-manifest.json'))
 
 /**
  * Convert the Next middleware name into a valid Edge Function name
@@ -73,13 +85,48 @@ const sanitizeName = (name: string) => `next_${name.replace(/\W/g, '_')}`
  * Initialization added to the top of the edge function bundle
  */
 const preamble = /* js */ `
-
-globalThis.process = { env: {...Deno.env.toObject(), NEXT_RUNTIME: 'edge', 'NEXT_PRIVATE_MINIMAL_MODE': '1' } }
-let _ENTRIES = {}
+import {
+  decode as _base64Decode,
+} from "https://deno.land/std@0.159.0/encoding/base64.ts";
 // Deno defines "window", but naughty libraries think this means it's a browser
 delete globalThis.window
-// Next uses "self" as a function-scoped global-like object
-const self = {}
+globalThis.process = { env: {...Deno.env.toObject(), NEXT_RUNTIME: 'edge', 'NEXT_PRIVATE_MINIMAL_MODE': '1' } }
+globalThis.EdgeRuntime = "netlify-edge"
+let _ENTRIES = {}
+
+// Next.js uses this extension to the Headers API implemented by Cloudflare workerd
+if(!('getAll' in Headers.prototype)) {
+  Headers.prototype.getAll = function getAll(name) {
+    name = name.toLowerCase();
+    if (name !== "set-cookie") {
+      throw new Error("Headers.getAll is only supported for Set-Cookie");
+    }
+    return [...this.entries()]
+      .filter(([key]) => key === name)
+      .map(([, value]) => value);
+  };
+}
+//  Next uses blob: urls to refer to local assets, so we need to intercept these
+const _fetch = globalThis.fetch
+const fetch = async (url, init) => {
+  try {
+    if (typeof url === 'object' && url.href?.startsWith('blob:')) {
+      const key = url.href.slice(5)
+      if (key in _ASSETS) {
+        return new Response(_base64Decode(_ASSETS[key]))
+      }
+    }
+    return await _fetch(url, init)
+  } catch (error) {
+    console.error(error)
+    throw error
+  }
+}
+
+// Next edge runtime uses "self" as a function-scoped global-like object, but some of the older polyfills expect it to equal globalThis
+// See https://nextjs.org/docs/basic-features/supported-browsers-features#polyfills
+const self = { ...globalThis, fetch }
+
 `
 
 // Slightly different spacing in different versions!
@@ -99,6 +146,22 @@ const getMiddlewareBundle = async ({
 }): Promise<string> => {
   const { publish } = netlifyConfig.build
   const chunks: Array<string> = [preamble]
+
+  if ('wasm' in edgeFunctionDefinition) {
+    for (const { name, filePath } of edgeFunctionDefinition.wasm) {
+      const wasm = await fs.readFile(join(publish, filePath))
+      chunks.push(`const ${name} = _base64Decode(${JSON.stringify(wasm.toString('base64'))}).buffer`)
+    }
+  }
+
+  if ('assets' in edgeFunctionDefinition) {
+    chunks.push(`const _ASSETS = {}`)
+    for (const { name, filePath } of edgeFunctionDefinition.assets) {
+      const wasm = await fs.readFile(join(publish, filePath))
+      chunks.push(`_ASSETS[${JSON.stringify(name)}] = ${JSON.stringify(wasm.toString('base64'))}`)
+    }
+  }
+
   for (const file of edgeFunctionDefinition.files) {
     const filePath = join(publish, file)
 
@@ -133,12 +196,18 @@ const writeEdgeFunction = async ({
   edgeFunctionDefinition,
   edgeFunctionRoot,
   netlifyConfig,
+  pageRegexMap,
+  appPathRoutesManifest = {},
   nextConfig,
+  cache,
 }: {
   edgeFunctionDefinition: EdgeFunctionDefinition
   edgeFunctionRoot: string
   netlifyConfig: NetlifyConfig
+  pageRegexMap?: Map<string, string>
+  appPathRoutesManifest?: Record<string, string>
   nextConfig: NextConfig
+  cache?: 'manual'
 }): Promise<
   Array<{
     function: string
@@ -179,12 +248,21 @@ const writeEdgeFunction = async ({
     matchers.push(...edgeFunctionDefinition.matchers)
   }
 
+  // If the EF matches a page, it's an app dir page so needs a matcher too
+  // The object will be empty if appDir isn't enabled in the Next config
+  if (pageRegexMap && edgeFunctionDefinition.page in appPathRoutesManifest) {
+    const regexp = pageRegexMap.get(appPathRoutesManifest[edgeFunctionDefinition.page])
+    if (regexp) {
+      matchers.push({ regexp })
+    }
+  }
+
   await writeJson(join(edgeFunctionDir, 'matchers.json'), matchers)
 
   // We add a defintion for each matching path
   return matchers.map((matcher) => {
     const pattern = stripLookahead(matcher.regexp)
-    return { function: name, pattern, name: edgeFunctionDefinition.name }
+    return { function: name, pattern, name: edgeFunctionDefinition.name, cache }
   })
 }
 export const cleanupEdgeFunctions = ({
@@ -217,7 +295,13 @@ export const writeDevEdgeFunction = async ({
 /**
  * Writes Edge Functions for the Next middleware
  */
-export const writeEdgeFunctions = async (netlifyConfig: NetlifyConfig) => {
+export const writeEdgeFunctions = async ({
+  netlifyConfig,
+  routesManifest,
+}: {
+  netlifyConfig: NetlifyConfig
+  routesManifest: RoutesManifest
+}) => {
   const manifest: FunctionManifest = {
     functions: [],
     version: 1,
@@ -229,6 +313,7 @@ export const writeEdgeFunctions = async (netlifyConfig: NetlifyConfig) => {
   const { publish } = netlifyConfig.build
   const nextConfigFile = await getRequiredServerFiles(publish)
   const nextConfig = nextConfigFile.config
+  const usesAppDir = nextConfig.experimental?.appDir
   await copy(getEdgeTemplatePath('../edge-shared'), join(edgeFunctionRoot, 'edge-shared'))
   await writeJSON(join(edgeFunctionRoot, 'edge-shared', 'nextConfig.json'), nextConfig)
 
@@ -258,13 +343,27 @@ export const writeEdgeFunctions = async (netlifyConfig: NetlifyConfig) => {
   // Older versions of the manifest format don't have the functions field
   // No, the version field was not incremented
   if (typeof middlewareManifest.functions === 'object') {
+    // When using the app dir, we also need to check if the EF matches a page
+    const appPathRoutesManifest = await loadAppPathRoutesManifest(netlifyConfig)
+
+    const pageRegexMap = new Map(
+      [...(routesManifest.dynamicRoutes || []), ...(routesManifest.staticRoutes || [])].map((route) => [
+        route.page,
+        route.regex,
+      ]),
+    )
+
     for (const edgeFunctionDefinition of Object.values(middlewareManifest.functions)) {
       usesEdge = true
       const functionDefinitions = await writeEdgeFunction({
         edgeFunctionDefinition,
         edgeFunctionRoot,
         netlifyConfig,
+        pageRegexMap,
+        appPathRoutesManifest,
         nextConfig,
+        // cache: "manual" is currently experimental, so we restrict it to sites that use experimental appDir
+        cache: usesAppDir ? 'manual' : undefined,
       })
       manifest.functions.push(...functionDefinitions)
     }
@@ -297,5 +396,3 @@ export const writeEdgeFunctions = async (netlifyConfig: NetlifyConfig) => {
 
   await writeJson(join(edgeFunctionRoot, 'manifest.json'), manifest)
 }
-
-/* eslint-enable max-lines */