+++ packages/runtime/src/helpers/redirects.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines */
 import type { NetlifyConfig } from '@netlify/build'
 import { yellowBright } from 'chalk'
 import { readJSON } from 'fs-extra'
@@ -19,6 +18,7 @@ import {
   isApiRoute,
   redirectsForNextRoute,
   redirectsForNextRouteWithData,
+  redirectsForNext404Route,
   routeToDataRoute,
 } from './utils'
 
@@ -187,6 +187,7 @@ const generateDynamicRewrites = ({
   basePath,
   buildId,
   i18n,
+  is404Isr,
 }: {
   dynamicRoutes: RoutesManifest['dynamicRoutes']
   prerenderedDynamicRoutes: PrerenderManifest['dynamicRoutes']
@@ -194,6 +195,7 @@ const generateDynamicRewrites = ({
   i18n: NextConfig['i18n']
   buildId: string
   middleware: Array<string>
+  is404Isr: boolean
 }): {
   dynamicRoutesThatMatchMiddleware: Array<string>
   dynamicRewrites: NetlifyConfig['redirects']
@@ -207,9 +209,11 @@ const generateDynamicRewrites = ({
     if (route.page in prerenderedDynamicRoutes) {
       if (matchesMiddleware(middleware, route.page)) {
         dynamicRoutesThatMatchMiddleware.push(route.page)
+      } else if (prerenderedDynamicRoutes[route.page].fallback === false && !is404Isr) {
+        dynamicRewrites.push(...redirectsForNext404Route({ route: route.page, buildId, basePath, i18n }))
       } else {
         dynamicRewrites.push(
-          ...redirectsForNextRoute({ buildId, route: route.page, basePath, to: ODB_FUNCTION_PATH, status: 200, i18n }),
+          ...redirectsForNextRoute({ route: route.page, buildId, basePath, to: ODB_FUNCTION_PATH, i18n }),
         )
       }
     } else {
@@ -263,6 +267,10 @@ export const generateRedirects = async ({
 
   const staticRouteEntries = Object.entries(prerenderedStaticRoutes)
 
+  const is404Isr = staticRouteEntries.some(
+    ([route, { initialRevalidateSeconds }]) => is404Route(route, i18n) && initialRevalidateSeconds !== false,
+  )
+
   const routesThatMatchMiddleware: Array<string> = []
 
   const { staticRoutePaths, staticIsrRewrites, staticIsrRoutesThatMatchMiddleware } = generateStaticIsrRewrites({
@@ -295,6 +303,7 @@ export const generateRedirects = async ({
     basePath,
     buildId,
     i18n,
+    is404Isr,
   })
   netlifyConfig.redirects.push(...dynamicRewrites)
   routesThatMatchMiddleware.push(...dynamicRoutesThatMatchMiddleware)
@@ -320,4 +329,3 @@ export const generateRedirects = async ({
     )
   }
 }
-/* eslint-enable max-lines */