+++ packages/runtime/src/templates/edge/bundle.js
@@ -2,6 +2,6 @@
  * This placeholder is replaced with the compiled Next.js bundle at build time
  * @param {Object} props
  * @param {import("./runtime.ts").RequestData} props.request
- * @returns {Promise<import("./utils.ts").FetchEventResult>}
+ * @returns {Promise<import("../edge-shared/utils.ts").FetchEventResult>}
  */
 export default async ({ request }) => {}